Grading Students
Every student receives a grade in the inclusive range from 0 to 100.

Any Grade less than 40 is a failing grade. Sam is a professor at the university and likes to round each student's according to these rules:

If the difference between the grade and the next multiple of 5 grade is less than 3 , round up to the next multiple of 5 .

- If the value of is less than 38 , no rounding occurs as the result will still be a failing grade.

- For example, 84 will be rounded to but will not be rounded because the rounding would result in a number that is less than .

Given the initial value of for each of Sam's students, write code to automate the rounding process. For each Sam's n student , round it according to the rules above and print the result on a new line**

n belongs To [1,60])

marks belongs To [1,100])