For loop
A for loop is a programming language statement which allows code to be repeatedly executed.

The syntax for this is

for ( <expression_1> ; <expression_2> ; <expression_3> )
    <statement>
expression_1 is used for intializing variables which are generally used for controlling terminating flag for the loop.
expression_2 is used to check for the terminating condition. If this evaluates to false, then the loop is terminated.
expression_3 is generally used to update the flags/variables.
A sample loop will be

for(int i = 0; i < 10; i++) {
    ...
}
Input Format

You will be given two positive integers a,b  and  (a<=b), separated by a newline.

Output Format

For each integer  (so all numbers in that range [a,b]):

If 1<=n && n<=9 , then print the English representation of it. That is "one" for 1, "two" for 2, and so on.
Else if n>9  and it is even, then print "even".
Else if  n>9 and it is odd, then print "odd".
Note:  represents the interval, i.e., 

Sample Input

8
11
S